device open close and reallocation is 100 percent, now moving on to the hph mnc code

+++ src/hph-ft260/hph-ft260.hpp
@@ -238,8 +238,8 @@ namespace hph
       void hid_api_check(void);
       void initialize_gpio(uint8_t device_count);
       void free_gpio(uint8_t device_count);
-      int open_device(uint8_t device_index);
-      void find_and_open_device(uint8_t device_index);
+      int open_device(int device_index);
+      void find_and_open_device(int device_index);
       void consolidate_used_memory(int ft260_devices);
 
       uint8_t i2c_data_report_id(uint8_t len);
